requirements.txt
click
httpx
python-dotenv
python-binance
binance